+++ source/d1image.h
@@ -6,7 +6,10 @@
 #include "d1gfx.h"
 #include "d1pal.h"
 
+// alpha value under which the color is considered as transparent
+#define COLOR_ALPHA_LIMIT 128
+
 class D1ImageFrame {
 public:
-    static bool load(D1GfxFrame &frame, QImage &image, bool clipped, D1Pal *pal);
+    static bool load(D1GfxFrame &frame, const QImage &image, bool clipped, D1Pal *pal);
 };