+++ source/levelcelview.cpp
@@ -9,6 +9,7 @@
 #include <QMessageBox>
 #include <QMimeData>
 
+#include "d1image.h"
 #include "mainwindow.h"
 #include "ui_levelcelview.h"
 
@@ -55,7 +56,7 @@ void LevelCelScene::dropEvent(QGraphicsSceneDragDropEvent *event)
         filePaths.append(url.toLocalFile());
     }
     // try to insert as frames
-    ((MainWindow *)this->view->window())->openImageFiles(filePaths, false);
+    ((MainWindow *)this->view->window())->openImageFiles(IMAGE_FILE_MODE::AUTO, filePaths, false);
 }
 
 void LevelCelScene::contextMenuEvent(QContextMenuEvent *event)
@@ -104,14 +105,18 @@ void LevelCelView::initialize(D1Gfx *g, D1Min *m, D1Til *t, D1Sol *s, D1Amp *a)
     this->til = t;
     this->sol = s;
     this->amp = a;
+    this->update();
+}
 
+void LevelCelView::update()
+{
     // Displaying CEL file path information
-    QFileInfo celFileInfo(this->gfx->getFilePath());
+    QFileInfo gfxFileInfo(this->gfx->getFilePath());
     QFileInfo minFileInfo(this->min->getFilePath());
     QFileInfo tilFileInfo(this->til->getFilePath());
     QFileInfo solFileInfo(this->sol->getFilePath());
     QFileInfo ampFileInfo(this->amp->getFilePath());
-    ui->celLabel->setText(celFileInfo.fileName() + ", " + minFileInfo.fileName() + ", " + tilFileInfo.fileName() + ", " + solFileInfo.fileName() + ", " + ampFileInfo.fileName());
+    ui->celLabel->setText(gfxFileInfo.fileName() + ", " + minFileInfo.fileName() + ", " + tilFileInfo.fileName() + ", " + solFileInfo.fileName() + ", " + ampFileInfo.fileName());
 
     ui->frameNumberEdit->setText(
         QString::number(this->gfx->getFrameCount()));
@@ -237,45 +242,304 @@ void LevelCelView::framePixelClicked(unsigned x, unsigned y)
     }
 }
 
-void LevelCelView::insertFrames(const QStringList &imagefilePaths, bool append)
+void LevelCelView::insertImageFiles(IMAGE_FILE_MODE mode, const QStringList &imagefilePaths, bool append)
+{
+    if (mode == IMAGE_FILE_MODE::FRAME || mode == IMAGE_FILE_MODE::AUTO) {
+        this->insertFrames(mode, imagefilePaths, append);
+    }
+    if (mode == IMAGE_FILE_MODE::SUBTILE || mode == IMAGE_FILE_MODE::AUTO) {
+        this->insertSubtiles(mode, imagefilePaths, append);
+    }
+    if (mode == IMAGE_FILE_MODE::TILE || mode == IMAGE_FILE_MODE::AUTO) {
+        this->insertTiles(mode, imagefilePaths, append);
+    }
+}
+
+void LevelCelView::assignFrames(const QImage &image, int subtileIndex, int frameIndex)
+{
+    QList<quint16> *frameIndices = nullptr;
+    if (subtileIndex >= 0) {
+        frameIndices = &this->min->getCelFrameIndices(subtileIndex);
+        frameIndices->clear();
+    }
+    // TODO: merge with D1Min::insertSubtile ?
+    QImage subImage = QImage(MICRO_WIDTH, MICRO_HEIGHT, QImage::Format_ARGB32);
+    for (int y = 0; y < image.height(); y += MICRO_HEIGHT) {
+        for (int x = 0; x < image.width(); x += MICRO_WIDTH) {
+            // subImage.fill(Qt::transparent);
+
+            bool hasColor = false;
+            for (int j = 0; j < MICRO_HEIGHT; j++) {
+                for (int i = 0; i < MICRO_WIDTH; i++) {
+                    const QColor color = image.pixelColor(x + i, y + j);
+                    if (color.alpha() >= COLOR_ALPHA_LIMIT) {
+                        hasColor = true;
+                    }
+                    subImage.setPixelColor(i, j, color);
+                }
+            }
+            if (frameIndices != nullptr) {
+                frameIndices->append(hasColor ? frameIndex + 1 : 0);
+            }
+            if (!hasColor) {
+                continue;
+            }
+
+            D1GfxFrame *frame = this->gfx->insertFrame(frameIndex, subImage);
+            LevelTabFrameWidget::selectFrameType(frame);
+            frameIndex++;
+        }
+    }
+}
+
+void LevelCelView::insertFrame(IMAGE_FILE_MODE mode, int index, const QString &imagefilePath)
+{
+    QImage image = QImage(imagefilePath);
+    if (image.isNull()) {
+        if (mode != IMAGE_FILE_MODE::AUTO) {
+            QMessageBox::critical(nullptr, "Error", "Failed open image file: " + imagefilePath);
+        }
+        return;
+    }
+
+    if ((image.width() % MICRO_WIDTH) != 0 || (image.height() % MICRO_HEIGHT) != 0) {
+        return;
+    }
+
+    if (mode == IMAGE_FILE_MODE::AUTO
+        && image.width() != MICRO_WIDTH && image.height() != MICRO_HEIGHT && image.width() != MICRO_WIDTH * EXPORT_LVLFRAMES_PER_LINE) {
+        // not a column of micros
+        // not a row or micros
+        // not a grouped micros from an export -> ignore
+        return;
+    }
+
+    this->assignFrames(image, -1, index);
+}
+
+void LevelCelView::insertFrames(IMAGE_FILE_MODE mode, const QStringList &imagefilePaths, bool append)
 {
     int prevFrameCount = this->gfx->getFrameCount();
 
     if (append) {
         // append the frame(s)
         for (int i = 0; i < imagefilePaths.count(); i++) {
-            int index = this->gfx->getFrameCount();
-            D1GfxFrame *frame = this->gfx->insertFrame(index, imagefilePaths[i]);
-            if (frame != nullptr) {
-                LevelTabFrameWidget::selectFrameType(frame);
-            }
+            this->insertFrame(mode, this->gfx->getFrameCount(), imagefilePaths[i]);
+        }
+        int deltaFrameCount = this->gfx->getFrameCount() - prevFrameCount;
+        if (deltaFrameCount == 0) {
+            return; // no new frame -> done
         }
         // jump to the first appended frame
-        int deltaFrameCount = this->gfx->getFrameCount() - prevFrameCount;
-        if (deltaFrameCount > 0) {
-            this->currentFrameIndex = prevFrameCount;
-        }
+        this->currentFrameIndex = prevFrameCount;
     } else {
         // insert the frame(s)
-        for (int i = 1; i <= imagefilePaths.count(); i++) {
-            D1GfxFrame *frame = this->gfx->insertFrame(this->currentFrameIndex, imagefilePaths[imagefilePaths.count() - i]);
-            if (frame != nullptr) {
-                LevelTabFrameWidget::selectFrameType(frame);
-            }
+        for (int i = imagefilePaths.count() - 1; i >= 0; i--) {
+            this->insertFrame(mode, this->currentFrameIndex, imagefilePaths[i]);
+        }
+        int deltaFrameCount = this->gfx->getFrameCount() - prevFrameCount;
+        if (deltaFrameCount == 0) {
+            return; // no new frame -> done
         }
         // shift references
-        int deltaFrameCount = this->gfx->getFrameCount() - prevFrameCount;
-        if (deltaFrameCount > 0) {
-            unsigned refIndex = this->currentFrameIndex + 1;
-            // shift frame indices of the subtiles
-            for (int i = 0; i < this->min->getSubtileCount(); i++) {
-                QList<quint16> &frameIndices = this->min->getCelFrameIndices(i);
-                for (int n = 0; n < frameIndices.count(); n++) {
-                    if (frameIndices[n] >= refIndex) {
-                        frameIndices[n] += deltaFrameCount;
-                    }
+        unsigned refIndex = this->currentFrameIndex + 1;
+        // shift frame indices of the subtiles
+        for (int i = 0; i < this->min->getSubtileCount(); i++) {
+            QList<quint16> &frameIndices = this->min->getCelFrameIndices(i);
+            for (int n = 0; n < frameIndices.count(); n++) {
+                if (frameIndices[n] >= refIndex) {
+                    frameIndices[n] += deltaFrameCount;
                 }
             }
+        }
+    }
+    // update the view
+    this->update();
+    this->displayFrame();
+}
+
+void LevelCelView::assignSubtiles(const QImage &image, int tileIndex, int subtileIndex)
+{
+    QList<quint16> *subtileIndices = nullptr;
+    if (tileIndex >= 0) {
+        subtileIndices = &this->til->getSubtileIndices(tileIndex);
+        subtileIndices->clear();
+    }
+    // TODO: merge with D1Til::insertTile ?
+    unsigned subtileWidth = this->min->getSubtileWidth() * MICRO_WIDTH;
+    unsigned subtileHeight = this->min->getSubtileHeight() * MICRO_HEIGHT;
+
+    QImage subImage = QImage(subtileWidth, subtileHeight, QImage::Format_ARGB32);
+    for (int y = 0; y < image.height(); y += subtileHeight) {
+        for (int x = 0; x < image.width(); x += subtileWidth) {
+            // subImage.fill(Qt::transparent);
+
+            bool hasColor = false;
+            for (unsigned j = 0; j < subtileHeight; j++) {
+                for (unsigned i = 0; i < subtileWidth; i++) {
+                    const QColor color = image.pixelColor(x + i, y + j);
+                    if (color.alpha() >= COLOR_ALPHA_LIMIT) {
+                        hasColor = true;
+                    }
+                    subImage.setPixelColor(i, j, color);
+                }
+            }
+
+            if (subtileIndices != nullptr) {
+                subtileIndices->append(subtileIndex);
+            } else if (!hasColor) {
+                continue;
+            }
+
+            this->min->insertSubtile(subtileIndex, subImage);
+            subtileIndex++;
+        }
+    }
+}
+
+void LevelCelView::insertSubtile(IMAGE_FILE_MODE mode, int index, const QString &imagefilePath)
+{
+    QImage image = QImage(imagefilePath);
+    if (image.isNull()) {
+        if (mode != IMAGE_FILE_MODE::AUTO) {
+            QMessageBox::critical(nullptr, "Error", "Failed open image file: " + imagefilePath);
+        }
+        return;
+    }
+
+    unsigned subtileWidth = this->min->getSubtileWidth() * MICRO_WIDTH;
+    unsigned subtileHeight = this->min->getSubtileHeight() * MICRO_HEIGHT;
+
+    if ((image.width() % subtileWidth) != 0 || (image.height() % subtileHeight) != 0) {
+        return;
+    }
+
+    if (mode == IMAGE_FILE_MODE::AUTO
+        && image.width() != subtileWidth && image.height() != subtileHeight && image.width() != subtileWidth * EXPORT_SUBTILES_PER_LINE) {
+        // not a column of subtiles
+        // not a row or subtiles
+        // not a grouped subtiles from an export -> ignore
+        return;
+    }
+
+    this->assignSubtiles(image, -1, index);
+}
+
+void LevelCelView::insertSubtiles(IMAGE_FILE_MODE mode, const QStringList &imagefilePaths, bool append)
+{
+    int prevSubtileCount = this->min->getSubtileCount();
+
+    if (append) {
+        // append the subtile(s)
+        for (int i = 0; i < imagefilePaths.count(); i++) {
+            this->insertSubtile(mode, this->min->getSubtileCount(), imagefilePaths[i]);
+        }
+        int deltaSubtileCount = this->min->getSubtileCount() - prevSubtileCount;
+        if (deltaSubtileCount == 0) {
+            return; // no new subtile -> done
+        }
+        // jump to the first appended subtile
+        this->currentSubtileIndex = prevSubtileCount;
+    } else {
+        // insert the subtile(s)
+        for (int i = imagefilePaths.count() - 1; i >= 0; i--) {
+            this->insertSubtile(mode, this->currentSubtileIndex, imagefilePaths[i]);
+        }
+        int deltaSubtileCount = this->min->getSubtileCount() - prevSubtileCount;
+        if (deltaSubtileCount == 0) {
+            return; // no new subtile -> done
+        }
+        // shift references
+        unsigned refIndex = this->currentSubtileIndex;
+        // shift subtile indices of the tiles
+        for (int i = 0; i < this->til->getTileCount(); i++) {
+            QList<quint16> &subtileIndices = this->til->getSubtileIndices(i);
+            for (int n = 0; n < subtileIndices.count(); n++) {
+                if (subtileIndices[n] >= refIndex) {
+                    subtileIndices[n] += deltaSubtileCount;
+                }
+            }
+        }
+    }
+    // update the view
+    this->update();
+    this->displayFrame();
+}
+
+void LevelCelView::insertTile(IMAGE_FILE_MODE mode, int index, const QString &imagefilePath)
+{
+    QImage image = QImage(imagefilePath);
+    if (image.isNull()) {
+        if (mode != IMAGE_FILE_MODE::AUTO) {
+            QMessageBox::critical(nullptr, "Error", "Failed open image file: " + imagefilePath);
+        }
+        return;
+    }
+
+    unsigned tileWidth = this->min->getSubtileWidth() * MICRO_WIDTH * TILE_WIDTH * TILE_HEIGHT;
+    unsigned tileHeight = this->min->getSubtileHeight() * MICRO_HEIGHT;
+
+    if ((image.width() % tileWidth) != 0 || (image.height() % tileHeight) != 0) {
+        return;
+    }
+
+    /*if (mode == IMAGE_FILE_MODE::AUTO
+        && image.width() != subtileWidth && image.height() != subtileHeight && image.width() != subtileWidth * EXPORT_TILES_PER_LINE) {
+        // not a column of tiles
+        // not a row or tiles
+        // not a grouped tiles from an export -> ignore
+        return;
+    }*/
+
+    QImage subImage = QImage(tileWidth, tileHeight, QImage::Format_ARGB32);
+    for (int y = 0; y < image.height(); y += tileHeight) {
+        for (int x = 0; x < image.width(); x += tileWidth) {
+            // subImage.fill(Qt::transparent);
+
+            bool hasColor = false;
+            for (unsigned j = 0; j < tileHeight; j++) {
+                for (unsigned i = 0; i < tileWidth; i++) {
+                    const QColor color = image.pixelColor(x + i, y + j);
+                    if (color.alpha() >= COLOR_ALPHA_LIMIT) {
+                        hasColor = true;
+                    }
+                    subImage.setPixelColor(i, j, color);
+                }
+            }
+
+            if (!hasColor) {
+                continue;
+            }
+
+            this->til->insertTile(index, subImage);
+            index++;
+        }
+    }
+}
+
+void LevelCelView::insertTiles(IMAGE_FILE_MODE mode, const QStringList &imagefilePaths, bool append)
+{
+    int prevTileCount = this->til->getTileCount();
+
+    if (append) {
+        // append the tile(s)
+        for (int i = 0; i < imagefilePaths.count(); i++) {
+            this->insertTile(mode, this->til->getTileCount(), imagefilePaths[i]);
+        }
+        int deltaTileCount = this->til->getTileCount() - prevTileCount;
+        if (deltaTileCount == 0) {
+            return; // no new tile -> done
+        }
+        // jump to the first appended tile
+        this->currentTileIndex = prevTileCount;
+    } else {
+        // insert the tile(s)
+        for (int i = imagefilePaths.count() - 1; i >= 0; i--) {
+            this->insertTile(mode, this->currentTileIndex, imagefilePaths[i]);
+        }
+        int deltaTileCount = this->til->getTileCount() - prevTileCount;
+        if (deltaTileCount == 0) {
+            return; // no new tile -> done
         }
     }
     // update the view
@@ -285,7 +549,19 @@ void LevelCelView::insertFrames(const QStringList &imagefilePaths, bool append)
 
 void LevelCelView::replaceCurrentFrame(const QString &imagefilePath)
 {
-    D1GfxFrame *frame = this->gfx->replaceFrame(this->currentFrameIndex, imagefilePath);
+    QImage image = QImage(imagefilePath);
+
+    if (image.isNull()) {
+        QMessageBox::critical(nullptr, "Error", "Failed open image file: " + imagefilePath);
+        return;
+    }
+
+    if (image.width() != MICRO_WIDTH || image.height() != MICRO_HEIGHT) {
+        QMessageBox::warning(this, "Warning", "The image must be 32px * 32px to be used as a level-frame.");
+        return;
+    }
+
+    D1GfxFrame *frame = this->gfx->replaceFrame(this->currentFrameIndex, image);
 
     if (frame != nullptr) {
         LevelTabFrameWidget::selectFrameType(frame);
@@ -359,6 +635,34 @@ void LevelCelView::cloneSubtile()
     this->displayFrame();
 }
 
+void LevelCelView::replaceCurrentSubtile(const QString &imagefilePath)
+{
+    QImage image = QImage(imagefilePath);
+
+    if (image.isNull()) {
+        QMessageBox::critical(nullptr, "Error", "Failed open image file: " + imagefilePath);
+        return;
+    }
+
+    unsigned subtileWidth = this->min->getSubtileWidth() * MICRO_WIDTH;
+    unsigned subtileHeight = this->min->getSubtileHeight() * MICRO_HEIGHT;
+
+    if (image.width() != subtileWidth || image.height() != subtileHeight) {
+        QMessageBox::warning(this, "Warning", "The image must be " + QString::number(subtileWidth) + "px * " + QString::number(subtileHeight) + "px to be used as a subtile.");
+        return;
+    }
+
+    int subtileIndex = this->currentSubtileIndex;
+    this->assignFrames(image, subtileIndex, this->gfx->getFrameCount());
+
+    // reset subtile flags
+    this->sol->setSubtileProperties(subtileIndex, 0);
+
+    // update the view
+    this->update();
+    this->displayFrame();
+}
+
 void LevelCelView::removeCurrentSubtile()
 {
     // check if the subtile is used
@@ -415,6 +719,34 @@ void LevelCelView::cloneTile()
     int cloneFrom = this->currentTileIndex;
     this->createTile();
     this->til->getSubtileIndices(this->currentTileIndex) = this->til->getSubtileIndices(cloneFrom);
+    this->displayFrame();
+}
+
+void LevelCelView::replaceCurrentTile(const QString &imagefilePath)
+{
+    QImage image = QImage(imagefilePath);
+
+    if (image.isNull()) {
+        QMessageBox::critical(nullptr, "Error", "Failed open image file: " + imagefilePath);
+        return;
+    }
+
+    unsigned tileWidth = this->min->getSubtileWidth() * MICRO_WIDTH * TILE_WIDTH * TILE_HEIGHT;
+    unsigned tileHeight = this->min->getSubtileHeight() * MICRO_HEIGHT;
+
+    if (image.width() != tileWidth || image.height() != tileHeight) {
+        QMessageBox::warning(this, "Warning", "The image must be " + QString::number(tileWidth) + "px * " + QString::number(tileHeight) + "px to be used as a tile.");
+        return;
+    }
+
+    int tileIndex = this->currentTileIndex;
+    this->assignSubtiles(image, tileIndex, this->min->getSubtileCount());
+
+    // reset tile flags
+    this->amp->setTileProperties(tileIndex, 0);
+
+    // update the view
+    this->update();
     this->displayFrame();
 }
 
@@ -539,66 +871,112 @@ void LevelCelView::ShowContextMenu(const QPoint &pos)
     MainWindow *mw = (MainWindow *)this->window();
 
     QMenu contextMenu(tr("Context menu"), this);
-    contextMenu.setToolTipsVisible(true);
 
-    QAction action0("Insert Frame", this);
+    QMenu frameMenu(tr("Frame"), this);
+    frameMenu.setToolTipsVisible(true);
+
+    QAction action0("Insert", this);
     action0.setToolTip("Add new frames before the current one");
     QObject::connect(&action0, SIGNAL(triggered()), mw, SLOT(on_actionInsert_Frame_triggered()));
-    contextMenu.addAction(&action0);
+    frameMenu.addAction(&action0);
 
-    QAction action1("Add Frame", this);
+    QAction action1("Add", this);
     action1.setToolTip("Add new frames at the end");
     QObject::connect(&action1, SIGNAL(triggered()), mw, SLOT(on_actionAdd_Frame_triggered()));
-    contextMenu.addAction(&action1);
+    frameMenu.addAction(&action1);
 
-    QAction action2("Replace Frame", this);
+    QAction action2("Replace", this);
     action2.setToolTip("Replace the current frame");
     QObject::connect(&action2, SIGNAL(triggered()), mw, SLOT(on_actionReplace_Frame_triggered()));
     if (this->gfx->getFrameCount() == 0) {
         action2.setEnabled(false);
     }
-    contextMenu.addAction(&action2);
+    frameMenu.addAction(&action2);
 
-    QAction action3("Del Frame", this);
+    QAction action3("Delete", this);
     action3.setToolTip("Delete the current frame");
     QObject::connect(&action3, SIGNAL(triggered()), mw, SLOT(on_actionDel_Frame_triggered()));
     if (this->gfx->getFrameCount() == 0) {
         action3.setEnabled(false);
     }
-    contextMenu.addAction(&action3);
+    frameMenu.addAction(&action3);
 
-    contextMenu.addSeparator();
+    contextMenu.addMenu(&frameMenu);
 
-    QAction action5("Create Subtile", this);
-    action5.setToolTip("Create a new subtile");
-    QObject::connect(&action5, SIGNAL(triggered()), mw, SLOT(on_actionCreate_Subtile_triggered()));
-    contextMenu.addAction(&action5);
+    QMenu subtileMenu(tr("Subtile"), this);
+    subtileMenu.setToolTipsVisible(true);
 
-    QAction action6("Delete Subtile", this);
-    action6.setToolTip("Delete the current subtile");
-    QObject::connect(&action6, SIGNAL(triggered()), mw, SLOT(on_actionDel_Subtile_triggered()));
-    if (this->min->getSubtileCount() == 0) {
-        action6.setEnabled(false);
-    }
-    contextMenu.addAction(&action6);
+    QAction action4("Create", this);
+    action4.setToolTip("Create a new subtile");
+    QObject::connect(&action4, SIGNAL(triggered()), mw, SLOT(on_actionCreate_Subtile_triggered()));
+    subtileMenu.addAction(&action4);
 
-    contextMenu.addSeparator();
+    QAction action5("Insert", this);
+    action5.setToolTip("Add new subtiles before the current one");
+    QObject::connect(&action5, SIGNAL(triggered()), mw, SLOT(on_actionInsert_Subtile_triggered()));
+    subtileMenu.addAction(&action5);
 
-    QAction action7("Create Tile", this);
-    action7.setToolTip("Create a new tile");
-    QObject::connect(&action7, SIGNAL(triggered()), mw, SLOT(on_actionCreate_Tile_triggered()));
+    QAction action6("Add", this);
+    action6.setToolTip("Add new subtiles at the end");
+    QObject::connect(&action6, SIGNAL(triggered()), mw, SLOT(on_actionAdd_Subtile_triggered()));
+    subtileMenu.addAction(&action6);
+
+    QAction action7("Replace", this);
+    action7.setToolTip("Replace the current subtile");
+    QObject::connect(&action7, SIGNAL(triggered()), mw, SLOT(on_actionReplace_Subtile_triggered()));
     if (this->min->getSubtileCount() == 0) {
         action7.setEnabled(false);
     }
-    contextMenu.addAction(&action7);
+    subtileMenu.addAction(&action7);
 
-    QAction action8("Delete Tile", this);
-    action8.setToolTip("Delete the current tile");
-    QObject::connect(&action8, SIGNAL(triggered()), mw, SLOT(on_actionDel_Tile_triggered()));
-    if (this->til->getTileCount() == 0) {
+    QAction action8("Delete", this);
+    action8.setToolTip("Delete the current subtile");
+    QObject::connect(&action8, SIGNAL(triggered()), mw, SLOT(on_actionDel_Subtile_triggered()));
+    if (this->min->getSubtileCount() == 0) {
         action8.setEnabled(false);
     }
-    contextMenu.addAction(&action8);
+    subtileMenu.addAction(&action8);
+
+    contextMenu.addMenu(&subtileMenu);
+
+    QMenu tileMenu(tr("Tile"), this);
+    tileMenu.setToolTipsVisible(true);
+
+    QAction action9("Create", this);
+    action9.setToolTip("Create a new tile");
+    QObject::connect(&action9, SIGNAL(triggered()), mw, SLOT(on_actionCreate_Tile_triggered()));
+    if (this->min->getSubtileCount() == 0) {
+        action9.setEnabled(false);
+    }
+    tileMenu.addAction(&action9);
+
+    QAction action10("Insert", this);
+    action10.setToolTip("Add new tiles before the current one");
+    QObject::connect(&action10, SIGNAL(triggered()), mw, SLOT(on_actionInsert_Tile_triggered()));
+    tileMenu.addAction(&action10);
+
+    QAction action11("Add", this);
+    action11.setToolTip("Add new tiles at the end");
+    QObject::connect(&action11, SIGNAL(triggered()), mw, SLOT(on_actionAdd_Tile_triggered()));
+    tileMenu.addAction(&action11);
+
+    QAction action12("Replace", this);
+    action12.setToolTip("Replace the current tile");
+    QObject::connect(&action12, SIGNAL(triggered()), mw, SLOT(on_actionReplace_Tile_triggered()));
+    if (this->til->getTileCount() == 0) {
+        action12.setEnabled(false);
+    }
+    tileMenu.addAction(&action12);
+
+    QAction action13("Delete", this);
+    action13.setToolTip("Delete the current tile");
+    QObject::connect(&action13, SIGNAL(triggered()), mw, SLOT(on_actionDel_Tile_triggered()));
+    if (this->til->getTileCount() == 0) {
+        action13.setEnabled(false);
+    }
+    tileMenu.addAction(&action13);
+
+    contextMenu.addMenu(&tileMenu);
 
     contextMenu.exec(mapToGlobal(pos));
 }
@@ -834,7 +1212,7 @@ void LevelCelView::dropEvent(QDropEvent *event)
         filePaths.append(url.toLocalFile());
     }
     // try to insert as frames
-    ((MainWindow *)this->window())->openImageFiles(filePaths, false);
+    ((MainWindow *)this->window())->openImageFiles(IMAGE_FILE_MODE::AUTO, filePaths, false);
 }
 
 void LevelCelView::on_addTileButton_clicked()